(BSR)[PRO] chore: fix a few imports, move DialogStockThingDeleteConfirm
