+++ pro/src/components/IndividualOffer/StocksThing/StocksThing.tsx
@@ -41,6 +41,7 @@ import strokeEuroIcon from '@/icons/stroke-euro.svg'
 import strokeFrancIcon from '@/icons/stroke-franc.svg'
 import { getSuccessMessage } from '@/pages/IndividualOffer/commons/getSuccessMessage'
 import { ActionBar } from '@/pages/IndividualOffer/components/ActionBar/ActionBar'
+import { DialogStockThingDeleteConfirm } from '@/pages/IndividualOffer/components/DialogStockThingDeleteConfirm/DialogStockThingDeleteConfirm'
 import { DatePicker } from '@/ui-kit/form/DatePicker/DatePicker'
 import { QuantityInput } from '@/ui-kit/form/QuantityInput/QuantityInput'
 import { TextInput } from '@/ui-kit/form/TextInput/TextInput'
@@ -49,7 +50,6 @@ import { ListIconButton } from '@/ui-kit/ListIconButton/ListIconButton'
 
 import { ActivationCodeFormDialog } from './ActivationCodeFormDialog/ActivationCodeFormDialog'
 import { STOCK_THING_FORM_DEFAULT_VALUES } from './constants'
-import { DialogStockThingDeleteConfirm } from './DialogStockThingDeleteConfirm'
 import styles from './StockThing.module.scss'
 import { submitToApi } from './submitToApi'
 import type { StockThingFormValues } from './types'
