+++ pro/src/pages/IndividualOffer/IndividualOfferPriceTable/components/PriceTableForm/PriceTableForm.tsx
@@ -7,13 +7,13 @@ import { assertOrFrontendError } from '@/commons/errors/assertOrFrontendError'
 import { isDateValid } from '@/commons/utils/date'
 import { getDepartmentCode } from '@/commons/utils/getDepartmentCode'
 import { toNumberOrNull } from '@/commons/utils/toNumberOrNull'
-import { DialogStockThingDeleteConfirm } from '@/components/IndividualOffer/DialogStockDeleteConfirm/DialogStockThingDeleteConfirm'
 import { ActivationCodeFormDialog } from '@/components/IndividualOffer/StocksThing/ActivationCodeFormDialog/ActivationCodeFormDialog'
 import fullCodeIcon from '@/icons/full-code.svg'
 import fulleMoreIcon from '@/icons/full-more.svg'
 import fullTrashIcon from '@/icons/full-trash.svg'
 import strokeEuroIcon from '@/icons/stroke-euro.svg'
 import strokeFrancIcon from '@/icons/stroke-franc.svg'
+import { DialogStockThingDeleteConfirm } from '@/pages/IndividualOffer/components/DialogStockThingDeleteConfirm/DialogStockThingDeleteConfirm'
 import { Button } from '@/ui-kit/Button/Button'
 import { ButtonVariant } from '@/ui-kit/Button/types'
 import { DatePicker } from '@/ui-kit/form/DatePicker/DatePicker'
